% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_formats.R
\name{convert_pop_to_long}
\alias{convert_pop_to_long}
\title{Convert from the \code{vip_population_3d} class \code{vip_population} class
(long format)}
\usage{
convert_pop_to_long(pop_3d)
}
\arguments{
\item{pop_3d}{An object of class \code{vip_population_3d}}
}
\value{
An object of class \code{vip_population}
}
\description{
Takes and object of class \code{vip_population_3d} and converts it to
the standard (long format) of S3 class \code{vip_population} (as
generated by \code{\link[vip:vip_population]{vip::vip_population()}}.
}
\seealso{
\code{\link[vip:vip_population]{vip::vip_population()}}, \code{\link[=convert_pop_to_3d]{convert_pop_to_3d()}}
}
\author{
Tini Garske
}
